refactor: extract balance detail view from wallet vc
