+++ Library/Constants.swift
@@ -1,0 +1,12 @@
+//
+//  Library
+//
+//  Created by 0 on 14.11.19.
+//  Copyright © 2019 Zap. All rights reserved.
+//
+
+import Foundation
+
+enum Constants {
+    static let buttonCornerRadius: CGFloat = 20
+}
