+++ Library/Scenes/Wallet/WalletViewModel.swift
@@ -14,11 +14,10 @@ import SwiftBTC
 final class WalletViewModel: NSObject {
     let lightningService: LightningService
     let shouldHideEmptyWalletState: Signal<Bool, Never>
-    let balanceSegments: [WalletBalanceSegment]
-    let circleGraphSegments = Observable<[CircleGraphView.Segment]>([])
-
-    var syncViewModel: SyncViewModel
-
+    
+    let syncViewModel: SyncViewModel
+    let balanceDetailViewModel: BalanceDetailViewModel
+    
     var network: Observable<Network?> {
         return lightningService.infoService.network
     }
@@ -36,12 +35,7 @@ final class WalletViewModel: NSObject {
         self.syncViewModel = SyncViewModel(lightningService: lightningService)
         
         let balanceService = lightningService.balanceService
-        
-        self.balanceSegments = [
-            WalletBalanceSegment(segment: .onChain, amount: balanceService.onChainConfirmed.toSignal()),
-            WalletBalanceSegment(segment: .lightning, amount: balanceService.lightningChannelBalance.toSignal()),
-            WalletBalanceSegment(segment: .pending, amount: balanceService.totalPending)
-        ]
+        self.balanceDetailViewModel = BalanceDetailViewModel(balanceService: balanceService)
         
         self.shouldHideEmptyWalletState = lightningService.balanceService.totalBalance
             .map { $0 > 0 }
@@ -56,18 +50,5 @@ final class WalletViewModel: NSObject {
             .distinctUntilChanged()
         
         super.init()
-        
-        combineLatest(balanceService.lightningChannelBalance, balanceService.onChainConfirmed, balanceService.totalPending)
-            .distinctUntilChanged { $0 != $1 }
-            .observeNext { [weak self] in
-                let (lightningBalance, onChainBalance, pendingBalance) = $0
-
-                self?.circleGraphSegments.value = [
-                    CircleGraphView.Segment(amount: onChainBalance, color: Segment.onChain.color),
-                    CircleGraphView.Segment(amount: lightningBalance, color: Segment.lightning.color),
-                    CircleGraphView.Segment(amount: pendingBalance, color: Segment.pending.color)
-                ]
-            }
-            .dispose(in: reactive.bag)
     }
 }
